requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
